% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_compile.R
\name{compile.link}
\alias{compile.link}
\alias{compile.link.file}
\title{Live-compile a C++ link plugin}
\usage{
compile.link(code, factory, name = factory, arguments = list(), cache = TRUE)

compile.link.file(
  file,
  factory,
  name = factory,
  arguments = list(),
  cache = TRUE
)
}
\arguments{
\item{code}{One character string containing complete C++ source.}

\item{factory}{Name of the exported factory defined by \code{code}.}

\item{name}{Friendly name retained as R metadata.}

\item{arguments}{Named list passed to the factory.}

\item{cache}{Reuse an identical compiled plugin in the current session.}

\item{file}{Path to a complete C++ source file.}
}
\value{
An owning external pointer returned by the factory.
}
\description{
Live-compile a C++ link plugin
}
\seealso{
Other live-compiled plugins:
\code{\link[=clear.plugin.cache]{clear.plugin.cache()}},
\code{\link[=compile.criterion]{compile.criterion()}},
\code{\link[=compile.family]{compile.family()}},
\code{\link[=compile.family.link]{compile.family.link()}},
\code{\link[=list.plugin.cache]{list.plugin.cache()}}
}
\concept{live-compiled plugins}
