% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egmifs-package.R
\docType{package}
\name{egmifs-package}
\alias{egmifs-package}
\title{egmifs: Extended generalized monotone incremental forward stagewise regression}
\description{
Extended generalized monotone incremental forward stagewise regression
as a modular R/Rcpp/C++ solution-path framework.
}
\author{
\strong{Maintainer}: Alikhan Anuarbekov \email{anuarali@fel.cvut.cz}

Authors:
\itemize{
  \item Alikhan Anuarbekov \email{anuarali@fel.cvut.cz}
  \item Jiří Kléma \email{klema@fel.cvut.cz}
}

}
\keyword{internal}
