% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.egmifs}
\alias{print.egmifs}
\title{Print an egmifs fit}
\usage{
\method{print}{egmifs}(x, digits = max(3L, getOption("digits") - 3L), max.criteria = 20L, ...)
}
\arguments{
\item{x}{An \code{egmifs} fit.}

\item{digits}{Number of significant digits.}

\item{max.criteria}{Maximum number of criterion selections to print.}

\item{...}{Unused.}
}
\description{
Print an egmifs fit
}
